Move appsync schema into appsync graphql api
