+++ aws/provider.go
@@ -328,7 +328,6 @@ func Provider() terraform.ResourceProvider {
 			"aws_appsync_api_key":                              resourceAwsAppsyncApiKey(),
 			"aws_appsync_datasource":                           resourceAwsAppsyncDatasource(),
 			"aws_appsync_graphql_api":                          resourceAwsAppsyncGraphqlApi(),
-			"aws_appsync_schema":                               resourceAwsAppsyncSchema(),
 			"aws_athena_database":                              resourceAwsAthenaDatabase(),
 			"aws_athena_named_query":                           resourceAwsAthenaNamedQuery(),
 			"aws_autoscaling_attachment":                       resourceAwsAutoscalingAttachment(),
