+++ website/docs/r/appsync_graphql_api.html.markdown
@@ -45,6 +45,22 @@ resource "aws_appsync_graphql_api" "example" {
 }
 ```
 
+### With Schema
+```hcl
+resource "aws_appsync_graphql_api" "example" {
+  authentication_type = "AWS_IAM"
+  name                = "example"
+  schema              = <<EOF
+schema {
+	query: Query
+}
+type Query {
+  test: Int
+}
+EOF
+}
+```
+
 ### OpenID Connect Authentication
 
 ```hcl
@@ -104,6 +120,7 @@ The following arguments are supported:
 * `log_config` - (Optional) Nested argument containing logging configuration. Defined below.
 * `openid_connect_config` - (Optional) Nested argument containing OpenID Connect configuration. Defined below.
 * `user_pool_config` - (Optional) The Amazon Cognito User Pool configuration. Defined below.
+* `schema` - (Required) The schema definition, in GraphQL schema language format
 
 ### log_config
